Rename invalidated variable spdx-copyright-name
spdx-copyright-name was introduced as a customizable variable and
consumed at one place in spdx.el. The identifier was later renamed as
spdx-copyright-prefix in a subsequent commit. However, the same change
was not made to the line where the variable is consumed. This leads to
failure of spdx-insert-copyright and spdx-insert-spdx-copyright
commands. This commit remedies that problem.